Return municipalities for users/{userId}
