+++ serverless/__tests__/api/users/getUser/integration.test.js
@@ -16,6 +16,7 @@ describe('getUser api test', () => {
     expect(json).toHaveProperty('user');
     expect(json.user).toHaveProperty('hasZipCode');
     expect(json.user).toHaveProperty('questions');
+    expect(json.user).toHaveProperty('municipalities');
 
     expect(response.status).toEqual(200);
   });
